Add a Playlist page
Adds a new page opened when the user selects a playlist. The page shows
info about the playlist (including overall duration), plus a list of all
of the videos that are part of the playlist.

+++ sfos/harbour-newpipe/src/mediainfo.cpp
@@ -198,8 +198,8 @@ QList<MediaInfo::MediaInfoSignal> MediaInfo::parseJsonChanges(QJsonObject const&
   QString name;
   QString uploaderName;
   QString category;
-  int viewCount;
-  int likeCount;
+  qint64 viewCount;
+  qint64 likeCount;
   QString content;
   QDateTime uploadDate;
   QString description;
